+++ memo/ChangeLog.md
@@ -14,6 +14,7 @@
 - main: support MSYS (motivated by SUCHMOKUO's report) `#D1264` 0000000
   - edit: support `\$` in `PS1` for MSYS2 `#D1265` 0000000
   - msys2: work around MSYS2 Bash bug of missing <kbd>CR</kbd> `#D1270` 0000000
+  - cygwin, msys2: support widget `paste-from-clipboard` `#D1271` 0000000
 
 ## Fix